pyproject.toml
[project]
name = "llm-inference-toolkit"
version = "0.1.0"
description = "Semantic response cache and context compression engine for LLM inference"
requires-python = ">=3.12"
dependencies = [
    "litellm>=1.50.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "redis>=5.2.0",
    "numpy>=2.1.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
inference-toolkit = "inference_toolkit.api.main:main"

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/inference_toolkit"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
